+++ packages/messagefactory/baseontology/builder.go
@@ -3,7 +3,7 @@ package baseontology
 import (
 	"github.com/iotaledger/goshimmer/packages/binary/tangle/model/message/payload"
 	"github.com/iotaledger/goshimmer/packages/binary/tangle/model/message/payload/data"
-	"github.com/iotaledger/goshimmer/packages/messsagefactory"
+	"github.com/iotaledger/goshimmer/packages/messagefactory"
 )
 
 type Builder struct {
@@ -12,7 +12,7 @@ type Builder struct {
 func (b *Builder) BuildPayload(raw []byte) *payload.Payload {
 	var dataPayload payload.Payload = data.New(raw)
 
-	messsagefactory.Events.PayloadConstructed.Trigger(&dataPayload)
+	messagefactory.Events.PayloadConstructed.Trigger(&dataPayload)
 
 	return &dataPayload
 }
